refactor: MainNavBar 코드 분리
